Moved logic for grabbing user's emails to user.js
